requirements.txt.txt
streamlit
torch
timm
pillow
albumentations
pandas
opencv-python-headless
